fixes to landing page
