+++ src/app/landing/landing.component.css
@@ -14,9 +14,13 @@
   border: 1.5px solid grey;
 }
 
+.slider {
+  color: var(--textColor);
+  background-color: var(--bgColor);
+}
+
 .release-notes {
   padding: 10px;
-  max-height: 100px;
   background-color: var(--bgColor);
 }
 
@@ -124,7 +128,7 @@ mat-large-icon-button {
   gap: 10px;
   align-items: center;
   justify-content: space-around;
-  background-color: #ffbf00ad;
+  background-color: #f8be0e;
   color: black;
   border: 2px solid black;
   border-radius: 20px;
@@ -171,9 +175,10 @@ mat-large-icon-button {
   gap: 10px;
   width: 350px;
   padding: 15px;
-  margin: 10px;
+  margin: 5px;
   box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
   max-width: 800px;
+  border: 1px solid var(--textColor);
 }
 
 .score-bug:hover {
@@ -225,14 +230,16 @@ span {
   font-weight: 700;
 }
 
+.toggle {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
 @media only screen and (max-width: 768px) {
   .notes-table {
     width: 95%;
     max-height: 375px;
-  }
-
-  .release-notes {
-    max-height: 125px;
   }
 
   .release-widget {
@@ -248,11 +255,13 @@ span {
     width: 95%;
     flex-direction: column;
     align-items: center;
+    align-self: center;
+    gap: 5px;
   }
 }
 
 @media (min-width: 1800px) {
   .twitter-feeds iframe {
-    height: 800px;
+    height: 600px;
   }
 }
