+++ src/app/landing/landing.component.html
@@ -4,6 +4,20 @@
   <span class="release-subheader">
     Your source for clean and colorful NHL analytics.
   </span>
+  <div class="toggle">
+    <mat-slide-toggle
+      [(ngModel)]="isDarkMode"
+      ngDefaultControl
+      labelPosition="before"
+      (change)="toggleDarkMode()">
+      <div style="display: flex; align-items: center; gap: 5px">
+        <span class="slider">Dark Mode</span>
+        <mat-icon class="slider">
+          {{ isDarkMode ? "nights_stay" : "wb_sunny" }}
+        </mat-icon>
+      </div>
+    </mat-slide-toggle>
+  </div>
   <div *ngIf="luckyPlayer$ | async as luckyPlayer" class="lucky-player-widget">
     <span class="lucky-subheader">Your Lucky Player is:</span>
     <a
